Move tree_hash from ssz into own crate
