+++ eth2/utils/tree_hash/src/resize.rs
